Introduce apt.py for apt related logic
